+++ src/Psl/Network/SocketInterface.php
@@ -15,9 +15,6 @@ interface SocketInterface extends IO\CloseHandleInterface
 {
     /**
      * Returns the address of the local side of the socket.
-     *
-     * @throws IO\Exception\AlreadyClosedException If the socket has already been closed.
-     * @throws Exception\RuntimeException If unable to retrieve local address.
      */
     public function getLocalAddress(): Address;
 }