+++ src/Psl/TCP/Internal/Listener.php
@@ -40,6 +40,8 @@ final class Listener implements TCP\ListenerInterface
      */
     private Channel\ReceiverInterface $receiver;
 
+    private readonly Network\Address $localAddress;
+
     /**
      * @param resource $impl
      * @param int<1, max> $idleConnections
@@ -47,6 +49,7 @@ final class Listener implements TCP\ListenerInterface
     public function __construct(mixed $impl, int $idleConnections = self::DEFAULT_IDLE_CONNECTIONS)
     {
         $this->impl = $impl;
+        $this->localAddress = Network\Internal\get_sock_name($impl);
 
         /**
          * @var Channel\ReceiverInterface<array{true, Stream}|array{false, Network\Exception\RuntimeException}> $receiver
@@ -120,11 +123,7 @@ final class Listener implements TCP\ListenerInterface
     #[Override]
     public function getLocalAddress(): Network\Address
     {
-        if (!is_resource($this->impl)) {
-            throw new Network\Exception\AlreadyStoppedException('Server socket has already been stopped.');
-        }
-
-        return Network\Internal\get_sock_name($this->impl);
+        return $this->localAddress;
     }
 
     /**