+++ src/Psl/UDP/Socket.php
@@ -39,6 +39,8 @@ final class Socket implements Network\SocketInterface, IO\StreamHandleInterface
      */
     private mixed $stream;
 
+    private readonly Network\Address $localAddress;
+
     /**
      * @param resource $stream
      */
@@ -46,6 +48,13 @@ final class Socket implements Network\SocketInterface, IO\StreamHandleInterface
     {
         $this->stream = $stream;
         stream_set_blocking($stream, false);
+
+        $name = @stream_socket_get_name($stream, false);
+        if ($name === false) {
+            throw new Network\Exception\RuntimeException('Failed to get local address.');
+        }
+
+        $this->localAddress = Internal\parse_address($name);
     }
 
     /**
@@ -211,22 +220,10 @@ final class Socket implements Network\SocketInterface, IO\StreamHandleInterface
         return [$data, Internal\parse_address($address)];
     }
 
-    /**
-     * Get the local address this socket is bound to.
-     *
-     * @throws Network\Exception\RuntimeException If unable to retrieve local address.
-     * @throws IO\Exception\AlreadyClosedException If the socket has already been closed.
-     */
     #[Override]
     public function getLocalAddress(): Network\Address
     {
-        $stream = $this->getResource();
-        $name = @stream_socket_get_name($stream, false);
-        if ($name === false) {
-            throw new Network\Exception\RuntimeException('Failed to get local address.');
-        }
-
-        return Internal\parse_address($name);
+        return $this->localAddress;
     }
 
     /**